+++ packages/styles/src/lib/colors.ts
@@ -42,6 +42,7 @@ const color = {
     red50: '#FF0000',
     red60: '#AE0000',
     sky40: '#01C0F4',
+    turquoise40: '#01CBD2',
     white: '#FFFFFF',
     yellow10: '#FEFAED',
     yellow20: '#F9EDBA',
@@ -52,7 +53,8 @@ const color = {
 export const gradientColors = {
     beauBlue:
         'radial-gradient(106.45% 108.64% at 32.33% -4.84%, #ECF5FC 0.52%, #CEE4F3 100%)',
-    blue: 'linear-gradient(113.54deg, rgba(60, 87, 140, 0.5) 14.91%, rgba(70, 86, 169, 0.5) 30.39%, rgba(129, 161, 225, 0.185) 55.76%), linear-gradient(151.07deg, #145559 33.25%, #414FA7 98.24%)',
+    blue:
+        'linear-gradient(113.54deg, rgba(60, 87, 140, 0.5) 14.91%, rgba(70, 86, 169, 0.5) 30.39%, rgba(129, 161, 225, 0.185) 55.76%), linear-gradient(151.07deg, #145559 33.25%, #414FA7 98.24%)',
     goldShiny:
         'linear-gradient(113.54deg, rgba(199, 148, 37, 0.6) 14.91%, rgba(204, 147, 23, 0.6) 14.92%, rgba(181, 147, 24, 0) 45.98%, rgba(250, 228, 30, 0) 55.76%, rgba(219, 200, 26, 0) 55.76%), linear-gradient(147.17deg, #DBBB14 48.73%, #B38311 98.22%)',
     lightBlue: